requirements.txt
pandas>=2.0.0
langgraph>=0.0.60
langchain>=0.1.0
langchain_qwq>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langsmith>=0.0.60
tavily-python>=0.3.0
wikipedia>=1.4.0
trustcall>=0.0.39
gradio>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
pyyaml>=6.0
redis>=4.5.0
